configure_network: Use routed forwarding mode instead of nat
By using routed forwarding mode, it is possible to access the guests
from the host without port forwarding.

This is used instead of the bridge mode to allow setting the domain for
hostname resolution using libvirt dnsmasq instance.

This also opens the ports used by the Keylime Agent on the firewall.

Signed-off-by: Anderson Toshiyuki Sasaki <ansasaki@redhat.com>